requirements.txt
python-telegram-bot
requests
aiohttp
networkx

tweepy
